added favs function to context!
