+++ src/App.jsx
@@ -1,18 +1,19 @@
 import './styles/styles.scss';
 import AppRouter from './routers/AppRouter';
-import { useState, createContext } from 'react';
+import { createContext } from 'react';
+import { toggleFavs } from './context/toggleFavs';
 
-// context for favorites object
+// context for favorites function
 export const FavContext = createContext();
 
 function App() {
   // context for favorites array state
-  const initialFavs = JSON.parse(localStorage.getItem('favMovieData')) || [];
-  const [favs, setFavs] = useState(initialFavs);
+  // const initialFavs = JSON.parse(localStorage.getItem('favMovieData')) || [];
+  // const [favs, setFavs] = useState(initialFavs);
 
   return (
     <>
-      <FavContext.Provider value={{ favs, setFavs }} >
+      <FavContext.Provider value={{ toggleFavs }} >
         <AppRouter />
       </FavContext.Provider>
     </>
